# notas_service/docs/swagger.yaml
basePath: /
definitions:
  models.Nota:
    properties:
      contenido:
        type: string
      id:
        type: string
      titulo:
        type: string
    type: object
  models.Usuario:
    properties:
      email:
        type: string
      id:
        type: string
      nombre:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API para gestionar notas - Guido Pisarra
  title: API de Notas
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Verifica las credenciales del usuario y devuelve token
      parameters:
      - description: Datos de inicio de sesión
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/models.Usuario'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Inicia sesión de usuario
      tags:
      - Usuarios
  /auth/registro:
    post:
      consumes:
      - application/json
      description: Crea un nuevo usuario en el sistema
      parameters:
      - description: Usuario a registrar
        in: body
        name: usuario
        required: true
        schema:
          $ref: '#/definitions/models.Usuario'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Registra un nuevo usuario
      tags:
      - Usuarios
  /notas:
    get:
      consumes:
      - application/json
      description: Obtiene una lista de todas las notas almacenadas
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Nota'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Obtiene todas las notas
      tags:
      - Notas
    post:
      consumes:
      - application/json
      description: Crea una nueva nota y la guarda en la base de datos
      parameters:
      - description: Nueva nota
        in: body
        name: nota
        required: true
        schema:
          $ref: '#/definitions/models.Nota'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Nota'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Crea una nueva nota
      tags:
      - Notas
  /notas/{id}:
    delete:
      consumes:
      - application/json
      description: Elimina una nota de la base de datos según el ID proporcionado
      parameters:
      - description: ID de la nota
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Elimina una nota
      tags:
      - Notas
    put:
      consumes:
      - application/json
      description: Actualiza los detalles de una nota específica en la base de datos
      parameters:
      - description: ID de la nota
        in: path
        name: id
        required: true
        type: string
      - description: Nota actualizada
        in: body
        name: nota
        required: true
        schema:
          $ref: '#/definitions/models.Nota'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Edita una nota existente
      tags:
      - Notas
swagger: "2.0"